romanNumeral_converter
This is my roman numeral converter I created. It can convert decimal values into roman numerals for numbers within the bounds of 1- 4999. It takes into account edge cases as well, which are numbers such as 4, 9, 40, and 90. The program does so by traversing two separate arrays, a decimal array, and a roman numeral array. It traverses the array, and then adds the values of the roman numerals onto an empty string.